init version

+++ Client/src/components/contanier/staff/Student/ViewStudent/index.js
@@ -1,5 +1,6 @@
 import React, { Component } from "react";
 import { toast } from "react-toastify";
+import { Redirect } from 'react-router-dom';
 
 import "./index.css";
 import Table from "../../../../common/Table";
@@ -12,34 +13,36 @@ export default class ViewStudent extends Component {
     headings: [],
     rows: [],
     search: '',
-    Error: {}
+    Error: {},
+    isloadding: true
   }
   handleInput = ({ target: { value, name } }) => this.setState({ [name]: value });
 
   handleDelete = (id) => {
     console.log(id);
-    fetch(`/api/v1/student/${id}`, {
-      method: 'DELETE',
-    })
-      .then(res => res.json())
-      .then(res => {
-        if (res.data) {
-          toast.success(res.data);
-          this.props.history.push('/staff/student/viewstudent');
-        }
-        else toast.error(res.error);
+    if (window.confirm('Delete the item?')) {
+      fetch(`/api/v1/student/${id}`, {
+        method: 'DELETE',
       })
+        .then(res => res.json())
+        .then(res => {
+          console.log(res);
+          if (res.data) {
+            toast.success(res.data);
+            this.props.history.push('/staff/student/viewstudent');
+          }
+          else toast.error(res.error);
+        })
+    }
   }
   handleSearch = () => {
-    if (this.state.search != '') {
-      //fillter the student
-      console.log('search')
-    }
-    else {
-      this.setState({ Error: {'search':'Enter vlaue'} })
-    }
+    /*console.log('object')
+    this.state.rows[0].filter(a =>
+
+      console.log(a.toString().search(/Fatma/))
+    )*/
   }
-  
+
   componentDidMount() {
     fetch('/api/v1/student', {
       method: 'GET'
@@ -50,24 +53,31 @@ export default class ViewStudent extends Component {
           const rows = res.data;
           let rowContent = [];
           rows.map(row => {
-            rowContent.push([row.id, row.firstname + ' ' + row.lastname, row.id, row.level, row.mobile_phone, <a onClick={() => this.handleDelete(row.id)}><i className="fa fa-trash delete-icon" aria-hidden="true" ></i></a>]);
+            rowContent.push([row.id, row.firstname + ' ' + row.lastname, row.id, row.level, row.mobile_phone,
+            <a onClick={() => this.handleDelete(row.id)}>
+              <i className="fa fa-trash delete-icon" aria-hidden="true" >
+              </i>
+            </a>
+            ]);
           })
-          this.state.headings = [
-            "Student Name",
-            "Student Number",
-            "Level ",
-            "Mobile Number",
-            ' '
-          ]
-          return this.setState({ rows: rowContent, Error: '' })
+          this.setState({
+            headings: [
+              "Student Name",
+              "Student Number",
+              "Level ",
+              "Mobile Number",
+              ' '
+            ]
+          })
+          return this.setState({ rows: rowContent, Error: '', isloadding: false })
         }
-        return this.setState({ Error: res.error })
+        return this.setState({ Error: res.error, isloadding: false })
       })
-      .catch(err => this.setState({ Error: 'Something happen error' }))
+      .catch(err => this.setState({ Error: 'Something happen error', isloadding: false }))
   }
 
   render() {
-    const { rows } = this.state;
+    const { rows, isloadding } = this.state;
     return (
       <>
         <h1 className="view-student-titel">
@@ -86,16 +96,24 @@ export default class ViewStudent extends Component {
             />
             <Button content="search" className="search-button" onClick={() => this.handleSearch()} />
           </div>
-          {rows ?
-            <>
-              <div className="result">Result:</div>
-              <div className="view-student">
-                <Table headings={this.state.headings} rows={this.state.rows} history={this.props.history} pathname={this.props.location.pathname} />
-              </div>
-            </>
-            :
-            !rows ? <h1>no coursesg</h1> : <Loading />
-          }
+
+
+          <div className="result">Result:</div>
+          <div className="view-student">
+            {isloadding ? <Loading /> : rows.length !== 0 ?
+              <Table headings={this.state.headings} rows={this.state.rows} history={this.props.history} pathname={this.props.location.pathname} />
+              :
+              <div className='no-student'>
+                There is no student untill now
+            <br />
+                Click <a
+                  className='add-student-link'
+                  onClick={() => this.props.history.push('/staff/student/addstudent')}
+                >
+                  here </a>
+                to add student</div>
+            }
+          </div>
         </div>
       </>
     );
